feat(redis): 支持 ZADD 命令及响应对象增强
- 新增 RedisResponse 的 setValue 方法和静态工厂方法
- 新增 ArrayResponse 和相关导入
- 修改 ZADD 命令 score 类型从 double 为 int
- 构造 ZADD 命令的 RESP 协议格式实现
- 添加 ArrayList 和 List 的相关导入支持命令参数构建

+++ src/main/java/tech/smartboot/redisun/ZAddCommand.java
@@ -6,5 +6,5 @@ package tech.smartboot.redisun;
  */
 public interface ZAddCommand {
 
-    boolean add(String key, double score, String member);
+    boolean add(String key, int score, String member);
 }
